Refactor  - some additions  - bring up to line with backend
